refactor: separate example contracts
