+++ packages/orchestration/src/proposals/start-stakeAtom.js
@@ -3,12 +3,12 @@ import { makeTracer } from '@agoric/internal';
 import { makeStorageNodeChild } from '@agoric/internal/src/lib-chainStorage.js';
 import { E } from '@endo/far';
 
-/** @import { StakeAtomSF,  StakeAtomTerms} from '../contracts/stakeAtom.contract' */
+/** @import { StakeAtomSF,  StakeAtomTerms} from '../examples/stakeAtom.contract' */
 
 const trace = makeTracer('StartStakeAtom', true);
 
 /**
- * @param {BootstrapPowers & { installation: {consume: {stakeAtom: Installation<import('../contracts/stakeAtom.contract.js').start>}}}} powers
+ * @param {BootstrapPowers & { installation: {consume: {stakeAtom: Installation<import('../examples/stakeAtom.contract.js').start>}}}} powers
  * @param {{options: StakeAtomTerms }} options
  */
 export const startStakeAtom = async (
